+++ src/shared/constants/global.constants.ts
@@ -13,6 +13,7 @@ export const SIGN_OUT_API_ENDPOINT = '/api/auth/sign-out'
 
 export const GET_QUOTE_API_ENDPOINT = '/api/quotes'
 export const MARGIN_PROFIT_API_ENDPOINT = '/api/margin-profit'
+export const ADDRESS_API_ENDPOINT = '/api/address'
 
 // Routes
 export const REGISTER_ROUTE = '/register';