+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## v0.24.0 (2025-11-30)
+
+### Pull Requests
+[#56](https://github.com/RafaelMoro/kraft-envios-fe/pull/56) | Add view addresses and fetch them
+
+
+## v0.23.0 (2025-11-30)
+
+### Pull Requests
+[#55](https://github.com/RafaelMoro/kraft-envios-fe/pull/55) | Add address subscreen and create address modal
+
+
 ## v0.22.2 (2025-11-29)
 
 ### Pull Requests